Clearer naming scheme
